+++ app/src/test/java/page/x/TestLRU.java
@@ -3,7 +3,7 @@ package page.x;
 import org.junit.jupiter.api.Test;
 
 import page.x.TLB.TLB;
-import page.x.entry.TlbEntry;
+import page.x.TLB.TlbEntry;
 import page.x.TLB.algoritmos.substituicao.LRU;
 import page.x.interruptions.MissInterruption;
 
@@ -14,13 +14,13 @@ import org.junit.jupiter.api.DisplayName;
 
 class TestLRU {
 
-    private final static int SIZE_LRU = 5;
+    private final static Long SIZE_LRU = 5L;
     private final static int QTD_PAIR = 10;
     private LRU<TlbEntry> lru1;
     private TLB tlb1;
     private TlbEntry[] entries;
     
-    void preencheTLB(TLB tlb, TlbEntry[] entries, Integer size) throws MissInterruption{
+    void preencheTLB(TLB tlb, TlbEntry[] entries, Long size) throws MissInterruption{
         for (int i = 0; i < size; i++) {
             tlb.addPaginaMapeada(entries[i].getVirtualPageNumber(), entries[i].getPageFrameNumber());
         }
@@ -87,7 +87,7 @@ class TestLRU {
     @DisplayName("Verificando a lógica da LRU-02")
     void preenchendoTLBeVerificandoLRU02() throws MissInterruption {
         this.preencheTLB(this.tlb1, this.entries, SIZE_LRU);
-        for (int i = SIZE_LRU; i < QTD_PAIR-1; i++) {
+        for (int i = SIZE_LRU.intValue(); i < QTD_PAIR-1; i++) {
             this.tlb1.addPaginaMapeada(entries[i].getVirtualPageNumber(), entries[i].getPageFrameNumber());
         }
         this.tlb1.mapearPagina(5L);
